[FIX/IMP] product_pack: Allow modify pack and do not reset after confirm website sale order (#201)
* [FIX] product_pack: Do not reset packs components when confirm sale order.

If a sale order is updated from the website whrn confirm the order we do
not expand the pack, we usit as it is.

* [ADD] product_pack: Allow modify pack from backend/website

Now we can choose if the product can be modified in the only int he backend or
if can be modified also in the frontend by the customers.

* [FIX] lint

+++ product_pack/models/sale_order_line.py
@@ -43,6 +43,8 @@ class SaleOrderLine(models.Model):
 
     @api.constrains('product_id', 'price_unit', 'product_uom_qty')
     def expand_pack_line(self):
+        if self._context.get('update_pricelist', False):
+            return
         detailed_packs = ['components_price', 'totalice_price', 'fixed_price']
         if (
                 self.state == 'draft' and
@@ -102,6 +104,8 @@ class SaleOrderLine(models.Model):
 
     @api.constrains('product_id')
     def expand_none_detailed_pack(self):
+        if self._context.get('update_pricelist', False):
+            return
         if self.product_id.pack_price_type == 'none_detailed_assited_price':
             # remove previus existing lines
             self.pack_line_ids.unlink()
@@ -137,7 +141,8 @@ class SaleOrderLine(models.Model):
         """ Do not let to edit a sale order line if this one belongs to pack
         """
         if self._origin.pack_parent_line_id and \
-           not self._origin.pack_parent_line_id.product_id.allow_modify_pack:
+           self._origin.pack_parent_line_id.product_id.allow_modify_pack \
+           not in ['only_backend', 'frontend_backend']:
             raise UserError(_(
                 'You can not change this line because is part of a pack'
                 ' included in this order'))
